+++ src/spotify/spotify.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { SpotifyService } from './spotify.service';
+import { SpotifyController } from './spotify.controller';
+
+@Module({
+  imports: [],
+  controllers: [SpotifyController],
+  providers: [SpotifyService],
+  exports: [SpotifyService]
+})
+export class SpotifyModule {}